Make return value of json_object_array_length an int
The function returns an int anyway, so it makes more sense to store it as this
instead of a size_t.  It also avoids a warning.

+++ libiview/parse_index.c
@@ -8,7 +8,7 @@ ssize_t iv_parse_index(const char *buf, size_t len IV_UNUSED,
         struct iv_series **index_ptr) {
     json_object *json_index, *json_element, *json_series;
     json_index = json_tokener_parse(buf);
-    const size_t index_len = json_object_array_length(json_index);
+    const int index_len = json_object_array_length(json_index);
     *index_ptr = (struct iv_series *)malloc(index_len * sizeof(struct iv_series));
     if(!*index_ptr) {
         return -IV_ENOMEM;
